Work on class Cell(decrease coupling). Add proxy DP
-Create class in the role of proxy: GameboardProxy. Use GameboardProxy
instead of Gameboard in GameEngine. Move the singleton pattern from
Gameboard to GameboardProxy. The aim was to be able to unit test
Gameboard, which was impossible if it was singleton.
- Decrease coupling of class Cell with Gameboard and GameServices by:
adding private fields and public properties IsMine, IsEmpty, IsDetonated
to class Cell; added method Detonate; using the properties and the
method instead of working directly with symbols; moving the constants
for field symbols from class Gameboard to class Cell